refactor : userPrice는 로또 번호와 다르게 분리하기위해 UserLottoPrice클래스로 분리
